+++ examples/experimental/assets_yaml_dsl/assets_yaml_dsl/domain_specific_dsl/user_scripts/fetch_the_tickers.py
@@ -1,4 +1,4 @@
-from dagster_ext import init_dagster_ext
+from dagster_pipes import init_dagster_ext
 
 context = init_dagster_ext()
