+++ examples/experimental/assets_yaml_dsl/assets_yaml_dsl/pure_assets_dsl/sql_script.py
@@ -1,6 +1,6 @@
 import sys
 
-from dagster_ext import init_dagster_ext
+from dagster_pipes import init_dagster_ext
 
 
 class SomeSqlClient:
