+++ python_modules/dagster-test/dagster_test/toys/external_execution/numbers_example/number_x.py
@@ -1,4 +1,4 @@
-from dagster_ext import init_dagster_ext
+from dagster_pipes import init_dagster_ext
 
 from .util import compute_data_version, store_asset_value
