+++ python_modules/dagster/dagster/_core/ext/client.py
@@ -2,7 +2,7 @@ from abc import ABC, abstractmethod
 from contextlib import contextmanager
 from typing import TYPE_CHECKING, Iterator, Optional
 
-from dagster_ext import (
+from dagster_pipes import (
     ExtContextData,
     ExtExtras,
     ExtParams,
