+++ python_modules/libraries/dagster-k8s/dagster_k8s/ext.py
@@ -25,7 +25,7 @@ from dagster._core.ext.utils import (
     ext_protocol,
     extract_message_or_forward_to_stdout,
 )
-from dagster_ext import (
+from dagster_pipes import (
     ExtDefaultMessageWriter,
     ExtExtras,
 )
